+++ extension.js
@@ -10,7 +10,6 @@ var GOOGLE = 'GOOGLE';
 var OUTLOOK = 'OUTLOOK';
 var OK = 'OK';
 var IGNORE = 'Snooze';
-var BUMP = "Bump";
 
 function isAllDay( parsedDateTime )
 {
@@ -137,10 +136,6 @@ function activate( context )
             {
                 acknowledgeNotification( event );
             }
-            else if( button == BUMP )
-            {
-                bumpEvent( event );
-            }
         }
 
         if( vscode.workspace.getConfiguration( 'calendar' ).get( 'showAllDayNotificationsAtStartup' ) )
@@ -159,11 +154,11 @@ function activate( context )
                             var label = isToday ? "Today" : "Tomorrow";
                             if( vscode.workspace.getConfiguration( 'calendar' ).get( 'stickyNotifications' ) )
                             {
-                                vscode.window.showErrorMessage( label + ": " + event.summary, OK, BUMP ).then( function( button ) { buttonClicked( button, event ); } );
+                                vscode.window.showErrorMessage( label + ": " + event.summary, OK ).then( function( button ) { buttonClicked( button, event ); } );
                             }
                             else
                             {
-                                vscode.window.showInformationMessage( label + ": " + event.summary, OK, BUMP ).then( function( button ) { buttonClicked( button, event ); } );
+                                vscode.window.showInformationMessage( label + ": " + event.summary, OK ).then( function( button ) { buttonClicked( button, event ); } );
                             }
                         }
                     }
@@ -194,10 +189,6 @@ function activate( context )
                     }
                 }
             }
-            else if( button === BUMP )
-            {
-                bumpEvent( event );
-            }
         }
 
         var config = vscode.workspace.getConfiguration( 'calendar' );
@@ -211,11 +202,11 @@ function activate( context )
 
             if( config.get( 'stickyNotifications' ) )
             {
-                vscode.window.showErrorMessage( text, OK, IGNORE, BUMP ).then( function( button ) { buttonClicked( button, event ); } );
+                vscode.window.showErrorMessage( text, OK, IGNORE ).then( function( button ) { buttonClicked( button, event ); } );
             }
             else
             {
-                vscode.window.showInformationMessage( text, OK, IGNORE, BUMP ).then( function( button ) { buttonClicked( button, event ); } );
+                vscode.window.showInformationMessage( text, OK, IGNORE ).then( function( button ) { buttonClicked( button, event ); } );
             }
         }
     }
@@ -262,42 +253,6 @@ function activate( context )
                 }
             } );
         }
-    }
-
-    function bumpEvent( event )
-    {
-        if( event.start.date )
-        {
-            event.start.date = utils.toISODate( new Date( event.start.date ).addDays( 1 ) );
-            event.end.date = utils.toISODate( new Date( event.end.date ).addDays( 1 ) );
-        }
-        if( event.start.dateTime )
-        {
-            event.start.dateTime = new Date( event.start.dateTime ).addDays( 1 ).toISOString();
-            event.end.dateTime = new Date( event.end.dateTime ).addDays( 1 ).toISOString();
-        }
-
-        googleCalendar.updateEvent( refresh, event );
-    }
-
-    function bumpEvent( node )
-    {
-        node = node ? node : selectedNode();
-
-        var event = node.event;
-
-        if( event.start.date )
-        {
-            event.start.date = utils.toISODate( new Date( event.start.date ).addDays( 1 ) );
-            event.end.date = utils.toISODate( new Date( event.end.date ).addDays( 1 ) );
-        }
-        if( event.start.dateTime )
-        {
-            event.start.dateTime = new Date( event.start.dateTime ).addDays( 1 ).toISOString();
-            event.end.dateTime = new Date( event.end.dateTime ).addDays( 1 ).toISOString();
-        }
-
-        googleCalendar.updateEvent( refresh, event );
     }
 
     function fetch()
@@ -795,7 +750,6 @@ function activate( context )
         context.subscriptions.push( vscode.commands.registerCommand( 'calendar.remove', remove ) );
         context.subscriptions.push( vscode.commands.registerCommand( 'calendar.setLocation', setLocation ) );
         context.subscriptions.push( vscode.commands.registerCommand( 'calendar.setReminder', setReminder ) );
-        context.subscriptions.push( vscode.commands.registerCommand( 'calendar.bumpEvent', bumpEvent ) );
 
         context.subscriptions.push( calendarViewExplorer.onDidExpandElement( function( e ) { calendarTree.setExpanded( e.element, true ); } ) );
         context.subscriptions.push( calendarView.onDidExpandElement( function( e ) { calendarTree.setExpanded( e.element, true ); } ) );